feat: 공통 함수들 util로 분리
